Fix equation bugs
