+++ marker/layout/layout.py
@@ -3,7 +3,7 @@ from typing import List
 from surya.layout import batch_layout_detection
 
 from marker.pdf.images import render_image
-from marker.pdf.utils import rescale_bbox
+from marker.schema.bbox import rescale_bbox
 from marker.schema.page import Page
 from marker.settings import settings
